+++ crates/teloxide-core/src/types/chat_full_info.rs
@@ -196,6 +196,11 @@ pub struct ChatFullInfoPublicChannel {
     /// identifier for a channel and vice versa.
     // SMELL: TBA uses here Integer instead of ChatId so we do that too :c
     pub linked_chat_id: Option<i64>,
+
+    /// `true`, if paid media messages can be sent or forwarded to the channel
+    /// chat. The field is available only for channel chats.
+    #[serde(default, skip_serializing_if = "std::ops::Not::not")]
+    pub can_send_paid_media: bool,
 }
 
 #[serde_with::skip_serializing_none]
@@ -697,6 +702,7 @@ mod tests {
                 kind: ChatFullInfoPublicKind::Channel(ChatFullInfoPublicChannel {
                     username: Some("channel_name".into()),
                     linked_chat_id: None,
+                    can_send_paid_media: false,
                 }),
                 description: None,
                 invite_link: None,