+++ crates/teloxide-core/CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add `InputPaidMedia`, `InputPaidMediaPhoto` and `InputPaidMediaVideo` structs
   - Add `sendPaidMedia` TBA method
   - Update documentation of the `copyMessage` and `copyMessages` methods
+  - Add `can_send_paid_media` field to `ChatFullInfoPublicChannel` struct
 
 ### Changed